requirements.txt
boto3
aws-cdk-lib==2.27.0 